Move entire return logic to another model
Due to a previous bad interpretation of solidus return flow creation we moved the
return creation from CustomerReturn to ReturnAuthorization model.
CustomerReturn will be implied when a shipwire return will be marked as
received.

ref https://github.com/nebulab/solidus_shipwire/pull/10/commits/a21632dc2ee620310888c0246a0a369f9ce4cd77
